Change type to fix tests

+++ include/osr/routing/profiles/foot.h
@@ -17,7 +17,7 @@ struct foot {
   static constexpr auto const kMaxMatchDistance = 100U;
 
   struct parameters {
-    double const speed_{IsWheelchair ? 0.8 : 1.2};
+    float const speed_{IsWheelchair ? 0.8F : 1.2F};
   };
 
   struct node {
